+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/runtime/dispatch/MapInterop.java
@@ -50,6 +50,7 @@ import com.oracle.truffle.espresso.vm.InterpreterToVM;
  * guest exception (in the case of unmodifiable map, for example).
  */
 @ExportLibrary(value = InteropLibrary.class, receiverType = StaticObject.class)
+@SuppressWarnings("truffle-abstract-export") // TODO GR-44080 Adopt BigInteger Interop
 public class MapInterop extends EspressoInterop {
     // region ### Hashes